# api/openapi.yaml
openapi: 3.0.3
info:
  title: Heromail API
  version: 0.1.0
  description: |
    Corporate mail service.

    This document is the source of truth for the HTTP contract. The Go server
    interface is generated from it, so an endpoint declared here that nobody
    implements fails to compile — the spec cannot quietly drift from the code.

    Two independent bearer credentials exist, and they are not interchangeable:

    * **session** — a signed-in person, issued by `/auth/login`. Guards `/auth/me`,
      `/auth/logout` and everything under `/workspaces`.
    * **apiKey** — a machine acting for one workspace, `hm_live_…`. Guards `/v1/*`.
servers:
  - url: https://api-dev-heromail.rnm.dev
    description: dev
  - url: http://localhost:8060
    description: local
...
security: []
paths:
...
  /auth/verify-email:
    post:
      tags: [auth]
      operationId: verifyEmail
      summary: Consume an emailed verification link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyEmailRequest' }
      responses:
        '200':
          description: Address verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '400':
          description: The token is unknown, expired or already used
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/resend-verification:
    post:
      tags: [auth]
      operationId: resendVerification
      summary: Send a fresh verification link
      description: |
        Always answers 202, whether or not the address is registered. Any other
        behaviour would be an account-enumeration oracle. Issuing a new link
        invalidates the previous one.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResendVerificationRequest' }
      responses:
        '202':
          description: Accepted, whether or not anything was sent
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/forgot-password:
    post:
      tags: [auth]
      operationId: forgotPassword
      summary: Send a password reset link
      description: |
        Always answers 202, whether or not the address is registered — any other
        behaviour would be an account-enumeration oracle. Issuing a new link
        invalidates the previous one.

        Accounts that sign in only through SSO have no password to reset and
        receive nothing.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '202':
          description: Accepted, whether or not anything was sent
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/reset-password:
    post:
      tags: [auth]
      operationId: resetPassword
      summary: Set a new password from a reset link
      description: |
        Consumes the emailed token, sets the new password, and **revokes every
        existing session** — a reset is what someone does when they believe the
        account is compromised, so leaving the attacker's session alive would
        defeat it.

        Reading the link also proves control of the address, so the account's
        email is marked verified if it was not already.

        The response is a fresh session for the browser that performed the
        reset, in the same shape as login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: Password changed; every other session is gone
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Credentials' }
        '400':
          description: The token is unknown, expired or already used, or the password is too weak
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: The user behind the current session
      security: [{ session: [] }]
      responses:
        '200':
          description: The signed-in user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /workspaces:
    get:
      tags: [workspaces]
      operationId: listWorkspaces
      summary: Workspaces the signed-in user belongs to
      security: [{ session: [] }]
      responses:
        '200':
          description: The user's workspaces, with their role in each
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [workspaces]
      operationId: createWorkspace
      summary: Create a workspace
      description: The creator becomes its owner.
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkspaceRequest' }
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: The slug is taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /workspaces/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Slug' }
    get:
      tags: [workspaces]
      operationId: getWorkspace
      summary: One workspace by slug
      description: |
        A workspace the user does not belong to is reported as absent, so
        membership cannot be probed by slug.
      security: [{ session: [] }]
      responses:
        '200':
          description: The workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/workspace:
    get:
      tags: [emails]
      operationId: getApiKeyWorkspace
      summary: The workspace the API key belongs to
      security: [{ apiKey: [] }]
      responses:
        '200':
          description: The workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /messages/search:
    get:
      tags: [emails]
      operationId: searchMessages
      summary: Search messages visible to the signed-in user
      description: Searches stored outbound messages across all workspaces the user belongs to.
      security: [{ session: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Matching messages, newest first
          content:
            application/json:
              schema:
                type: object
                required: [messages]
                properties:
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/Email' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/emails:
    post:
      tags: [emails]
      operationId: sendEmail
      summary: Queue a message for sending
      description: |
        **Asynchronous.** The message is recorded and handed to the queue; the
        request does not wait for the mail transport. Poll `GET /v1/emails/{id}`
        for the outcome, or read the event trail.

        This is why there is no `502` here any more: at the time we answer,
        nobody has talked to a mail server yet. A transport failure surfaces as
        `status: failed` on the message, and the worker retries with backoff
        until it gives up and marks it `dead`.
      security: [{ apiKey: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Scoped to the workspace. Repeating a request with the same key
            returns the original message and does not send again. Omitting it
            means no deduplication at all.
          schema: { type: string, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendEmailRequest' }
      responses:
        '202':
          description: |
            Queued. For a repeated idempotency key this returns the original
            message unchanged, whatever state it has reached by now.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SendEmailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/attachments:
    post:
      tags: [emails]
      operationId: uploadAttachment
      summary: Upload a file to attach to a later message
      description: |
        Stores the file in object storage and returns an id. Reference that id
        in `SendEmailRequest.attachments` to send it; an upload that is never
        referenced is an orphan and goes nowhere.

        `503` means the deployment has no bucket configured yet — attachments
        are wired end to end before credentials exist, this is not a bug.
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Stored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Attachment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413':
          description: The file, or the multipart body, is too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: Attachment storage is not configured on this deployment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/emails/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [emails]
      operationId: getEmail
      summary: One message
      description: |
        Scoped to the API key's workspace. Another workspace's message is
        reported as absent.
      security: [{ apiKey: [] }]
      responses:
        '200':
          description: The message
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Email' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      description: Session token from `/auth/register` or `/auth/login`.
    apiKey:
      type: http
      scheme: bearer
      description: Workspace API key, `hm_live_…`, minted by the issue-api-key command.
  responses:
    BadRequest:
      description: The request is malformed or fails validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, malformed, expired or revoked credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource within the caller's scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: |
        The caller is over a rate limit. `Retry-After` carries the number of
        seconds until the window resets.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the limit resets.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        The single error envelope for the whole API. `/v1/*` used to return a
        bare `{"error":"text"}`, which the frontend client could not read.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable. Branch on this, not on the message.
              example: validation_failed
            message:
              type: string
              description: Human-readable, safe to show a user.
            details:
              description: Optional per-field or per-code extra context.
...
    Slug:
      type: string
      pattern: '^[a-z0-9][a-z0-9-]{1,61}[a-z0-9]$'
      description: Lowercase tenant handle used in URLs.
      example: acme
...
    UserEnvelope:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/User' }
    Credentials:
      type: object
      required: [token, user, expiresAt]
      properties:
        token:
          type: string
          description: Session bearer token. Shown once; not recoverable.
        user: { $ref: '#/components/schemas/User' }
        expiresAt: { type: string, format: date-time }
...
    VerifyEmailRequest:
      type: object
      required: [token]
      additionalProperties: false
      properties:
        token: { type: string }
    ForgotPasswordRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, password]
      additionalProperties: false
      properties:
        token: { type: string }
        password:
          type: string
          minLength: 10
          maxLength: 72
          description: Same rules as registration.
    ResendVerificationRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
    Workspace:
      type: object
      required: [id, slug, name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        slug: { $ref: '#/components/schemas/Slug' }
        name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    WorkspaceList:
      type: object
      required: [workspaces]
      properties:
        workspaces:
          type: array
          items: { $ref: '#/components/schemas/WorkspaceMembership' }
    CreateWorkspaceRequest:
      type: object
      required: [name, slug]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        slug: { $ref: '#/components/schemas/Slug' }
...
    Email:
      type: object
      required: [id, workspace_id, from, to, subject, status, attempts, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        from: { type: string, format: email }
        to:
          type: array
          items: { type: string, format: email }
        subject: { type: string }
        html: { type: string, nullable: true }
        text: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/EmailStatus' }
        provider_message_id: { type: string, nullable: true }
        idempotency_key: { type: string, nullable: true }
        attempts: { type: integer }
        last_error: { type: string, nullable: true }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/Attachment' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }